requirements.txt
streamlit
pandas
requests
python-dotenv
yfinance
plotly

